% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\docType{data}
\name{idmc}
\alias{idmc}
\title{IDMC Data}
\format{
A tibble.
\describe{
\item{year}{The year (e.g. 2020).}
\item{coo_name}{Country of origin name.}
\item{coo}{Country of origin UNHCR code.}
\item{coo_iso}{Country of origin ISO code.}
\item{coa_name}{Country of asylum name.}
\item{coa}{Country of asylum UNHCR code.}
\item{coa_iso}{Country of asylum ISO code.}
\item{total}{The total number of internally displaced persons due to conflict and violence.}
}
}
\source{
\href{https://www.unhcr.org/refugee-statistics/}{UNHCR Refugee Data Finder}.
}
\usage{
idmc
}
\description{
Data on Internally displaced persons due to conflict and violence
that is produced by the \href{https://www.internal-displacement.org/}{Internal Displacement Monitoring Centre}.
}
\details{
See the notes on the data structure in the
\href{https://www.unhcr.org/refugee-statistics/methodology/}{methodology pages} for more details.
}
\examples{
idmc
}
\keyword{datasets}
