% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\docType{data}
\name{solutions}
\alias{solutions}
\title{Solutions}
\format{
A tibble.
\describe{
\item{year}{The year (e.g. 2020).}
\item{coo_name}{Country of origin name.}
\item{coo}{Country of origin UNHCR code.}
\item{coo_iso}{Country of origin ISO code.}
\item{coa_name}{Country of asylum name.}
\item{coa}{Country of asylum UNHCR code.}
\item{coa_iso}{Country of asylum ISO code.}
\item{returned_refugees}{The number of refugee returnees.}
\item{resettlement}{The number of resettlement arrivals.}
\item{naturalisation}{The number of refugees that have naturalised.}
\item{returned_idps}{The number of IDP returnees.}
}
}
\source{
\href{https://www.unhcr.org/refugee-statistics/}{UNHCR Refugee Data Finder}.
}
\usage{
solutions
}
\description{
Data on solutions record those refugees and IDPs that have availed a durable solution.
}
\details{
See the notes on the data structure in the
\href{https://www.unhcr.org/refugee-statistics/methodology/}{methodology pages} for more details.
}
\examples{
solutions
}
\keyword{datasets}
