% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries}
\format{
A tibble.
\describe{
\item{iso_code}{ISO3 Country code.}
\item{unhcr_code}{UNHCR Country code.}
\item{name}{UNSD Name.}
\item{unhcr_region}{UNHCR Region name.}
\item{unsd_region}{UNSD Region name.}
\item{unsd_subregion}{UNSD Sub-region name.}
\item{unsd_imregion}{SDG Region name.}
\item{sdg_region}{SDG Region name.}
}
}
\source{
\href{https://www.unhcr.org/refugee-statistics/}{UNHCR Refugee Data Finder}.
}
\usage{
countries
}
\description{
Country codes, names, UN major areas and UNHCR regional bureaux/operations.
}
\examples{
countries
}
\keyword{datasets}
